fixed MT components
