+++ src/index.css
@@ -1,9 +1,12 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   font-weight: 400;
   color: rgba(255, 255, 255, 0.87);
-  background: #1a1a1a;
-
+  /* background: #1a1a1a; */
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
