+++ src/main.jsx
@@ -1,14 +1,16 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
-import { HashRouter } from 'react-router-dom'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App.jsx";
+import "./index.css";
+import { HashRouter } from "react-router-dom";
+import { ThemeProvider } from "@material-tailwind/react";
 
-ReactDOM.createRoot(document.getElementById('root')).render(
+ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <HashRouter  >
-    <App />
+    <HashRouter>
+      <ThemeProvider>
+        <App />
+      </ThemeProvider>
     </HashRouter>
-    
-  </React.StrictMode>,
-)
+  </React.StrictMode>
+);
