+++ src/pages/login-page/LoginCard.jsx
@@ -1,85 +1,199 @@
+import React from "react";
+
 import {
   Card,
+  CardHeader,
+  CardBody,
   Input,
-  Checkbox,
   Button,
   Typography,
+  Tabs,
+  TabsHeader,
+  TabsBody,
+  Tab,
+  TabPanel,
+  Select,
+  Option,
 } from "@material-tailwind/react";
- 
-export function LoginCard() {
+import {
+  BanknotesIcon,
+  CreditCardIcon,
+  LockClosedIcon,
+} from "@heroicons/react/24/solid";
+import { SiFireship } from "react-icons/si";
+function formatCardNumber(value) {
+  const val = value.replace(/\s+/g, "").replace(/[^0-9]/gi, "");
+  const matches = val.match(/\d{4,16}/g);
+  const match = (matches && matches[0]) || "";
+  const parts = [];
+
+  for (let i = 0, len = match.length; i < len; i += 4) {
+    parts.push(match.substring(i, i + 4));
+  }
+
+  if (parts.length) {
+    return parts.join(" ");
+  } else {
+    return value;
+  }
+}
+
+function formatExpires(value) {
+  return value
+    .replace(/[^0-9]/g, "")
+    .replace(/^([2-9])$/g, "0$1")
+    .replace(/^(1{1})([3-9]{1})$/g, "0$1/$2")
+    .replace(/^0{1,}/g, "0")
+    .replace(/^([0-1]{1}[0-9]{1})([0-9]{1,2}).*/g, "$1/$2");
+}
+
+export default function LoginCard() {
+  const [type, setType] = React.useState("card");
+  const [cardNumber, setCardNumber] = React.useState("");
+  const [cardExpires, setCardExpires] = React.useState("");
+
   return (
-     <Card color="transparent" shadow={false}>
-      <Typography variant="h4" color="blue-gray">
-        Sign Up
-      </Typography>
-      <Typography color="gray" className="mt-1 font-normal">
-        Nice to meet you! Enter your details to register.
-      </Typography>
-      <form className="mt-8 mb-2 w-80 max-w-screen-lg sm:w-96">
-        <div className="mb-1 flex flex-col gap-6">
-          <Typography variant="h6" color="blue-gray" className="-mb-3">
-            Your Name
-          </Typography>
-          <Input
-            size="lg"
-            placeholder="name@mail.com"
-            className=" !border-t-blue-gray-200 focus:!border-t-gray-900"
-            labelProps={{
-              className: "before:content-none after:content-none",
-            }}
-          />
-          <Typography variant="h6" color="blue-gray" className="-mb-3">
-            Your Email
-          </Typography>
-          <Input
-            size="lg"
-            placeholder="name@mail.com"
-            className=" !border-t-blue-gray-200 focus:!border-t-gray-900"
-            labelProps={{
-              className: "before:content-none after:content-none",
-            }}
-          />
-          <Typography variant="h6" color="blue-gray" className="-mb-3">
-            Password
-          </Typography>
-          <Input
-            type="password"
-            size="lg"
-            placeholder="********"
-            className=" !border-t-blue-gray-200 focus:!border-t-gray-900"
-            labelProps={{
-              className: "before:content-none after:content-none",
-            }}
-          />
+    <Card className="w-full max-w-[24rem]">
+      <CardHeader
+        color="gray"
+        floated={false}
+        shadow={false}
+        className="m-0 grid place-items-center px-4 py-8 text-center"
+      >
+        <div className="mb-4 h-20 p-6 text-white">
+          <SiFireship size={50}/>
         </div>
-        <Checkbox
-          label={
-            <Typography
-              variant="small"
-              color="gray"
-              className="flex items-center font-normal"
-            >
-              I agree the
-              <a
-                href="#"
-                className="font-medium transition-colors hover:text-gray-900"
-              >
-                &nbsp;Terms and Conditions
-              </a>
-            </Typography>
-          }
-          containerProps={{ className: "-ml-2.5" }}
-        />
-        <Button className="mt-6" fullWidth>
-          sign up
-        </Button>
-        <Typography color="gray" className="mt-4 text-center font-normal">
-          Already have an account?{" "}
-          <a href="#" className="font-medium text-gray-900">
-            Sign In
-          </a>
+        <Typography variant="h4" color="white">
+          Ignite your own project , or help someone else begin one!
         </Typography>
-      </form>
+      </CardHeader>
+      <CardBody>
+        <Tabs value={type} className="overflow-visible">
+          <TabsHeader className="relative z-0 ">
+            <Tab value="card" onClick={() => setType("card")}>
+              Sign In
+            </Tab>
+            <Tab value="paypal" onClick={() => setType("paypal")}>
+              Sign Up
+            </Tab>
+          </TabsHeader>
+          <TabsBody
+            className="!overflow-x-hidden !overflow-y-visible"
+            animate={{
+              initial: {
+                x: type === "card" ? 400 : -400,
+              },
+              mount: {
+                x: 0,
+              },
+              unmount: {
+                x: type === "card" ? 400 : -400,
+              },
+            }}
+          >
+            <TabPanel value="card" className="p-0">
+              <form className="mt-12 flex flex-col gap-4">
+                <div>
+                  <Typography
+                    variant="small"
+                    color="blue-gray"
+                    className="mb-2 font-medium"
+                  >
+                    Username
+                  </Typography>
+                  <Input
+                    type="email"
+                    placeholder="name@mail.com"
+                    className=" !border-t-blue-gray-200 focus:!border-t-gray-900"
+                    labelProps={{
+                      className: "before:content-none after:content-none",
+                    }}
+                  />
+                </div>
+
+                <div className="my-3">
+                  <Typography
+                    variant="small"
+                    color="blue-gray"
+                    className="mb-2 font-medium "
+                  >
+                    Password
+                  </Typography>
+
+                  <Input maxLength={19} type="password" />
+                </div>
+                <Button size="lg">Login</Button>
+                <Typography
+                  variant="small"
+                  color="gray"
+                  className="mt-2 flex items-center justify-center gap-2 font-medium opacity-60"
+                >
+                  <LockClosedIcon className="-mt-0.5 h-4 w-4" /> Payments are
+                  secure and encrypted
+                </Typography>
+              </form>
+            </TabPanel>
+            <TabPanel value="paypal" className="p-0">
+              <form className="mt-12 flex flex-col gap-4">
+                <div>
+                  <Typography
+                    variant="small"
+                    color="blue-gray"
+                    className="mb-2 font-medium"
+                  >
+                    Username
+                  </Typography>
+                  <Input
+                    type="text"
+                    placeholder="name@mail.com"
+                    className=" !border-t-blue-gray-200 focus:!border-t-gray-900"
+                    labelProps={{
+                      className: "before:content-none after:content-none",
+                    }}
+                  />
+                </div>
+                <div>
+                  <Typography
+                    variant="small"
+                    color="blue-gray"
+                    className="mb-2 font-medium"
+                  >
+                    Email
+                  </Typography>
+                  <Input
+                    type="email"
+                    placeholder="name@mail.com"
+                    className=" !border-t-blue-gray-200 focus:!border-t-gray-900"
+                    labelProps={{
+                      className: "before:content-none after:content-none",
+                    }}
+                  />
+                </div>
+                <div className="my-3">
+                  <Typography
+                    variant="small"
+                    color="blue-gray"
+                    className="mb-2 font-medium "
+                  >
+                    Password
+                  </Typography>
+
+                  <Input maxLength={19} type="password" />
+                </div>
+                <Button size="lg">Login</Button>
+                <Typography
+                  variant="small"
+                  color="gray"
+                  className="mt-2 flex items-center justify-center gap-2 font-medium opacity-60"
+                >
+                  <LockClosedIcon className="-mt-0.5 h-4 w-4" /> Payments are
+                  secure and encrypted
+                </Typography>
+              </form>
+            </TabPanel>
+          </TabsBody>
+        </Tabs>
+      </CardBody>
     </Card>
   );
 }
