+++ src/pages/login-page/LoginPage.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import ConnectWalletButton from "../../assets/components/ConnectWalletButton";
-import { LoginCard } from "./LoginCard";
 import Web3 from "web3";
+import LoginCard from "./LoginCard";
+import styled from "@emotion/styled";
 const LoginPage = () => {
   const [loading, setLoading] = useState(false);
   const [address, setAddress] = useState("");
@@ -27,17 +27,15 @@ const LoginPage = () => {
 
   const onPressLogout = () => setAddress("");
   return (
-    <div>
-      {/* <div>
-        <ConnectWalletButton
-          onPressConnect={onPressConnect}
-          onPressLogout={onPressLogout}
-          loading={loading}
-          address={address}
-        />
-      </div> */}
-      <LoginCard/>
-    </div>
+    <StyledDiv>
+      <LoginCard />
+    </StyledDiv>
   );
 };
+const StyledDiv = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-self: center;
+  align-items: center;
+`;
 export default LoginPage;
